+++ plugins/wasm-go/extensions/ai-proxy/README.md
@@ -108,7 +108,7 @@ Ollama 所对应的 `type` 为 `ollama`。它特有的配置字段如下：
 
 #### 混元
 
-混元 所对应的 `type` 为 `hunyuan`。它特有的配置字段如下：
+混元所对应的 `type` 为 `hunyuan`。它特有的配置字段如下：
 
 | 名称                | 数据类型   | 填写要求 | 默认值 | 描述                                           |
 |-------------------|--------|------|-----|----------------------------------------------|
@@ -556,7 +556,7 @@ provider:
   "stop_reason": "end_turn"
 }
 ```
-### 使用 OpenAI 协议代理 混元 服务
+### 使用 OpenAI 协议代理混元服务
 
 **配置信息**
