+++ plugins/wasm-go/extensions/ai-proxy/provider/provider.go
@@ -124,10 +124,12 @@ type ProviderConfig struct {
 	// @Title zh-CN Ollama Server Port
 	// @Description zh-CN 仅适用于 Ollama 服务。Ollama 服务器的端口号。
 	ollamaServerPort uint32 `required:"false" yaml:"ollamaServerPort" json:"ollamaServerPort"`
-	// @Title zh-CN hunyuan api's key/id for authorization
-	// @Description zh-CN 仅适用于Hun Yuan AI服务。API key/id 参考：https://cloud.tencent.com/document/api/1729/101843#Golang
+	// @Title zh-CN hunyuan api key for authorization
+	// @Description zh-CN 仅适用于Hun Yuan AI服务鉴权，API key/id 参考：https://cloud.tencent.com/document/api/1729/101843#Golang
 	hunyuanAuthKey string `required:"false" yaml:"hunyuanAuthKey" json:"hunyuanAuthKey"`
-	hunyuanAuthId  string `required:"false" yaml:"hunyuanAuthId" json:"hunyuanAuthId"`
+	// @Title zh-CN hunyuan api id for authorization
+	// @Description zh-CN 仅适用于Hun Yuan AI服务鉴权
+	hunyuanAuthId string `required:"false" yaml:"hunyuanAuthId" json:"hunyuanAuthId"`
 	// @Title zh-CN 模型名称映射表
 	// @Description zh-CN 用于将请求中的模型名称映射为目标AI服务商支持的模型名称。支持通过“*”来配置全局映射
 	modelMapping map[string]string `required:"false" yaml:"modelMapping" json:"modelMapping"`
@@ -194,7 +196,6 @@ func (c *ProviderConfig) Validate() error {
 
 	if c.typ == "" {
 		return errors.New("missing type in provider config")
-
 	}
 	initializer, has := providerInitializers[c.typ]
 	if !has {
